requirements.txt
django==4.1.0
pillow==10.2.0
django-braces==1.15.0
django-embed-video==1.4.4
pymemcache==3.5.2
djangorestframework==3.13.1